ui: edit character side-by-side with chat, change ChatOptions layout
- Added "Chara" menu item to chat options.
- Due to the now large number of ChatOptions menu item, I've changed that menu's layout a little.
- "Chara" menu item opens the character editor in a right-side pane which turns into a modal on small screens.
- Character editor has been reworked to be more beautiful.
