+++ web/pages/Settings/UISettings.tsx
@@ -230,6 +230,7 @@ const UISettings: Component = () => {
           msg={toBotMsg(bot, '*I wave excitedly* Hello world!\nHow are you today?', { _id: '1' })}
           onRemove={noop}
           sendMessage={() => {}}
+          editingChar={false}
         />
 
         <Show when={state.profile}>
@@ -243,6 +244,7 @@ const UISettings: Component = () => {
             })}
             onRemove={noop}
             sendMessage={() => {}}
+            editingChar={false}
           />
         </Show>
       </div>
