+++ web/shared/Modal.tsx
@@ -30,7 +30,7 @@ const Modal: Component<Props> = (props) => {
   const minHeight = createMemo(() => (props.fixedHeight ? 'modal-height-fixed' : ''))
 
   const defaultSubmit = (ev: Event) => {
-    ev.preventDefault()
+    ev.preventDefault
   }
 
   return (
